+++ Tests/TestSocketIO/SocketParseErrorTest.swift
@@ -83,7 +83,6 @@ class SocketParseErrorTest: XCTestCase {
 private class FakeParseErrorEngine: SocketEngineSpec {
     weak var client: SocketEngineClient?
     private(set) var closed = false
-    private(set) var compress = false
     private(set) var connected = false
     var connectParams: [String: Any]? = nil
     private(set) var cookies: [HTTPCookie]? = nil
@@ -98,7 +97,6 @@ private class FakeParseErrorEngine: SocketEngineSpec {
     private(set) var socketPath = ""
     private(set) var urlPolling = URL(string: "http://localhost/")!
     private(set) var urlWebSocket = URL(string: "http://localhost/")!
-    private(set) var websocket = false
 
 
     private(set) var disconnectReasons = [String]()